DROID-3739 fix
